requirements.txt
scrapling[all]
google-genai
python-dotenv
colorama
plyer
patchright
